practice/.claude/skills/pdf-to-deck/requirements.txt
python-pptx>=0.6.23
pypdf>=4.0.0
requests>=2.31.0
